Rename NPaaceConfig
