+++ src/main/java/com/skycatdev/nopeacefuldespawn/mixin/AllowInPeacefulMixinSilverFishFix.java
@@ -15,7 +15,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import java.util.Objects;
 import java.util.Random;
 
-import static com.skycatdev.nopeacefuldespawn.NPaaceConfig.PEACEFUL_SPAWNERS;
+import static com.skycatdev.nopeacefuldespawn.NoPeacefulDespawnConfig.PEACEFUL_SPAWNERS;
 
 @Mixin(InfestedBlock.class)
 public abstract class AllowInPeacefulMixinSilverFishFix {
