+++ src/main/java/com/skycatdev/nopeacefuldespawn/mixin/HostileEntityMixin.java
@@ -10,7 +10,7 @@ import org.spongepowered.asm.mixin.injection.At;
 
 import java.util.Objects;
 
-import static com.skycatdev.nopeacefuldespawn.NPaaceConfig.PEACEFUL_SPAWNERS;
+import static com.skycatdev.nopeacefuldespawn.NoPeacefulDespawnConfig.PEACEFUL_SPAWNERS;
 
 @Mixin(HostileEntity.class)
 public abstract class HostileEntityMixin {
